pyproject.toml
[project]
name = "live-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.2.1",
    "aioftp>=0.27.2",
    "aiohttp>=3.13.1",
    "arrow>=1.4.0",
    "chonkie>=1.4.0",
    "crawl4ai>=0.7.6",
    "ddgs>=9.5.2",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pgvector>=0.4.1",
    "playwright>=1.55.0",
    "playwright-stealth>=2.0.0",
    "pre-commit>=4.4.0",
    "prefect>=3.4.25",
    "psycopg>=3.2.12",
    "pypdf>=6.1.3",
    "python-dotenv>=1.1.1",
    "reportlab>=4.4.4",
    "ruff>=0.14.4",
    "sqlmodel>=0.0.27",
    "tenacity>=9.1.2",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "transformers>=4.57.1",
]
